Split base.Syntax in different traits
